+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Prevent changes to AzureMachinePool Spot VM configuration after creation.
+
 ## [2.0.1] - 2021-01-27
 
 ### Fixed